add werrror (all warnings are errors) and fix those warnings

+++ Backend/data/database_connector.hpp
@@ -7,7 +7,7 @@ class dataBaseStart{
         sql::Connection *connection = mysql_driver->connect("tcp://127.0.0.1:3306", "root", "mypass");
         sql::Statement *statement;
         sql::ResultSet *result_set;
-        std::string createStatement(std::vector<std::string> input, std::string table, int tableSize = 0);
+        std::string createStatement(std::vector<std::string> input, std::string table, size_t tableSize = 0);
         std::vector<std::pair<std::string,std::string>> readQuestions();
         void updateQuestions();
     public:
